+++ blueman/plugins/applet/NMDUNSupport.py
@@ -8,7 +8,7 @@ from blueman.plugins.AppletPlugin import AppletPlugin
 import dbus
 from gi.repository import GLib
 from blueman.gui.Notification import Notification
-from blueman.Sdp import uuid128_to_uuid16, DIALUP_NET_SVCLASS_ID
+from blueman.Sdp import DIALUP_NET_SVCLASS_ID
 from blueman.Functions import get_icon, composite_icon
 import weakref
 import logging
@@ -101,7 +101,7 @@ class NMDUNSupport(AppletPlugin):
         pass
 
     def rfcomm_connect_handler(self, service, reply, err):
-        if DIALUP_NET_SVCLASS_ID == uuid128_to_uuid16(service.uuid):
+        if DIALUP_NET_SVCLASS_ID == service.short_uuid:
             ConnectionHandler(self, service, reply, err)
             return True
         else: